feat(AIService): implement cascading updates for AI credentials and enhance model validation logic
